updated tests for stages
