+++ pages/having-a-baby.js
@@ -3,7 +3,7 @@ import { useEffect, useRef, useContext, useReducer } from "react";
 import { LanguageContext } from "../context/languageProvider";
 import Select from "../components/atoms/Select";
 import Card from "../components/atoms/Card";
-import Stages2 from "../components/atoms/Stagesv2";
+import Stages from "../components/atoms/Stages";
 import { getLocalJourneys } from "./api/getData";
 
 import en from "../locales/en";
@@ -108,7 +108,7 @@ export default function lifejourney({ journeysData }) {
             </div>
 
             <div className="mt-4 w-full">
-              <Stages2
+              <Stages
                 journeys={journeysData}
                 selectPlaceholder={t.stagesSelect}
                 labelText={t.stagesLabel}
